Change types

+++ src/generator/headers/outputAux.h
@@ -3,7 +3,7 @@
 
 #include <string>
 
-std::string writePoint(double , double , double);
+std::string writePoint(float , float , float);
 void dumpFile(const std::ostringstream&, const std::string&);
 
 #endif
